Created oh-video-videojs player child component. It is lazy-loaded and provides decoupling of external player dependency from widget.
