+++ bundles/org.openhab.ui/web/src/components/widgets/system/oh-video.vue
@@ -1,27 +1,28 @@
 <template>
   <div class="player">
-    <video ref="videoPlayer" class="video-js vjs-fluid">
-      Sorry, your browser doesn't support embedded videos.
-    </video>
+    <oh-video-videojs
+      :src="this.src"
+      :type="this.config.type"
+      :startManually="this.config.startManually"
+      :hideControls="this.config.hideControls"
+    />
   </div>
 </template>
 
-<style lang="stylus"></style>
-
 <script>
-import videojs from "video.js";
-import "video.js/dist/video-js.css";
 import mixin from "../widget-mixin";
 import { OhVideoDefinition } from "@/assets/definitions/widgets/system";
 
 export default {
   mixins: [mixin],
   widget: OhVideoDefinition,
+  components: {
+    'oh-video-videojs': () => import(/* webpackChunkName: "oh-video-videojs" */ './oh-video-videojs.vue')
+  },
   data() {
     return {
       t: this.$utils.id(),
       src: null,
-      player: null,
     };
   },
   watch: {
@@ -41,50 +42,18 @@ export default {
     },
   },
   mounted() {
-    console.log("Mounted oh-video");
     if (this.config.item) {
       this.loadItemURL();
     } else {
       this.src = this.config.url;
-      this.createPlayer()
-    }
-  },
-  beforeDestroy() {
-    if (this.player) {
-      this.player.dispose();
     }
   },
   methods: {
-    createPlayer() {
-      console.log("Creating player for source: " + this.src )
-      if (this.player) {
-        this.player.dispose();
-      }
-      const playerOpts = {
-        liveui: true,
-        autoplay: this.config.startManually ? false : "muted",
-        controls: !this.config.hideControls,
-        sources: [
-          {
-            src: this.src,
-            type: this.config.type,
-          },
-        ],
-      }
-      this.player = videojs(
-        this.$refs.videoPlayer,
-        playerOpts,
-        function onPlayerReady() {
-          console.log("onPlayerReady", this);
-        }
-      );
-    },
     loadItemURL() {
       this.$oh.api
         .getPlain(`/rest/items/${this.config.item}/state`, "text/plain")
         .then((data) => {
           this.src = data;
-          this.createPlayer()
         });
     },
   },
